+++ lib/sigv4.dart
@@ -1,1 +1,3 @@
-export 'src/signature.dart';
+export 'src/sigv4.dart';
+export 'src/client.dart';
+export 'src/base_client.dart';