+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Changelog
+
+## 2.0.0
+
+BREAKING CHANGES:
+
+- Removed Sigv4Request, only Sigv4Client is needed
+- Exposed methods for building headers and canonical path independently
+- Exposed a wrapper method that returns a `http` Request object
+
+- Added documentation and README
+
 ## 1.0.0
 
-* Initial version
+- Initial version